r3/process-creation-posix: snprintf format-specifier warning.
svn:sync-xref-src-repo-rev: r91947

+++ src/VBox/Runtime/r3/posix/process-creation-posix.cpp
@@ -1,4 +1,4 @@
-/* $Id: process-creation-posix.cpp 49530 2013-11-18 13:08:36Z noreply@oracle.com $ */
+/* $Id: process-creation-posix.cpp 50292 2014-01-30 16:30:46Z ramshankar.venkataraman@oracle.com $ */
 /** @file
  * IPRT - Process Creation, POSIX.
  */
@@ -240,7 +240,7 @@ static void rtSolarisContractPostForkParent(int templateFd, pid_t pid)
     /* Abandon this contract we just created. */
     char ctlPath[PATH_MAX];
     size_t len = snprintf(ctlPath, sizeof(ctlPath),
-                          CTFS_ROOT "/process/%d/ctl", ctId);
+                          CTFS_ROOT "/process/%ld/ctl", (long)ctId);
     if (len >= sizeof(ctlPath))
         return;
     int ctlFd = open64(ctlPath, O_WRONLY);
